Rename CacheTypeChecker to CacheTagChecker
This renaming reflects a clearer and more accurate purpose for the interface and its implementations. Updated references and dependencies throughout the codebase to align with the new name. This improves code clarity and consistency.
